refactor: 커플 연결 UI 컴포넌트 역할별로 세분화
