+++ src/views/connect/model/types.ts
@@ -1,0 +1,2 @@
+export type ConnectStepType = 'create' | 'code'
+export type ConnectStep = 'date' | 'nickname' | 'create-code' | 'insert-code' | 'complete'
